+++ Program.cs
@@ -7,14 +7,14 @@ namespace LargestNumber
         
         public static void Main(string[] args)
         {
-            int First, Second, Third;
+            double First, Second, Third;
             Console.WriteLine("Largest number of given numbers!");
             GetInput get = new GetInput();
-            First = get.GetInt();
-            Second = get.GetInt();
-            Third = get.GetInt();
+            First = get.GetDouble();
+            Second = get.GetDouble();
+            Third = get.GetDouble();
             CompareValues LN = new CompareValues();
-            int largest = LN.LargestNum(First, Second, Third);
+            double largest = LN.LargestNum(First, Second, Third);
             Console.WriteLine("Largest number of {0} {1} and {2} is : {3}", First, Second, Third, largest);
         }
     }
